cims/requirements.txt
APScheduler==3.9.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
cssselect2==0.7.0
Django==4.1.1
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
django-daraja==1.3.0
django-dbbackup==4.0.2
django-environ==0.11.2
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-scheduler==0.10.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
fonttools==4.39.3
html5lib==1.1
icalendar==4.1.0
idna==3.4
lxml==4.9.3
oscrypto==1.3.0
phonenumbers==8.13.23
Pillow==9.5.0
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.6.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.15.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==4.7.1
tzdata==2022.4
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
weasyprint==58.1
webencodings==0.5.1
xhtml2pdf==0.2.11
xlwt==1.3.0
zopfli==0.2.2
